refactor(recovery): extract comment state module and facade
